Reverted change to elements check
It's preferable to have the option of passing `false` or an empty array for `elements`. This way we can decide whether to keep previous suggestions visible (since empty array evals to true in JS) or remove them (i.e. by passing `false` or similar falsey var as the value for `elements`).